Nomis dataset content types

Nomis content type metadata is included in annotation tags, in the form of contenttype/<contenttype> in the annotationtitle column in the annotations.annotation list-column returned from nomis_data_info(). For example, the content types returned from dataset "NM_1658_1", using nomis_data_info("NM_1658_1"), are "geoglevel", "2001census" and "sources".

nomis_content_type(content_type, id = NULL)

Arguments

content_type
A string with the content type to return metadata on.

id
A string with an optional content_type id.

Value

A tibble with metadata on a given content type.

See also

Examples

# \donttest{
a <- nomis_content_type("sources")

tibble::glimpse(a)

#> Rows: 16
#> Columns: 6
#> $ description <chr> "A count of home Civil Service employees. It excludes the…
#> $ id          <chr> "acses", "aps", "ashe", "bres", "census", "cc", "dwp", "v…
#> $ name        <chr> "Annual Civil Service Employment Survey", "Annual Populat…
#> $ type        <chr> "DataSource", "DataSource", "DataSource", "DataSource", "…
#> $ item        <list> [NULL, NULL, NULL, NULL, <data.frame[4 x 5]>, NULL, <dat…
#> $ status      <chr> NA, NA, NA, NA, NA, NA, NA, "archived", NA, NA, NA, NA, N…

b <- nomis_content_type("sources", id = "census")

tibble::glimpse(b)

#> Rows: 4
#> Columns: 9
#> $ description      <chr> "Census data from the 1981, 1991, 2001 and 2011 Cens…
#> $ id               <chr> "census", "census", "census", "census"
#> $ item.description <chr> "The 2011 Census was taken on 27th March 2011. The f…
#> $ item.id          <chr> "census_2011", "census_2001", "census_1991", "census…
#> $ item.item        <list> [<data.frame[13 x 5]>, <data.frame[6 x 5]>, NULL, N…
#> $ item.name        <chr> "2011 Census", "2001 Census", "1991 Census", "1981 C…
#> $ item.type        <chr> "DataSourceSubGroup", "DataSourceSubGroup", "DataSou…
#> $ name             <chr> "Census of Population", "Census of Population", "Cen…
#> $ type             <chr> "DataSource", "DataSource", "DataSource", "DataSourc…

# }