Nomis dataset content types

Nomis content type metadata is included in annotation tags, in the form of contenttype/<contenttype> in the annotationtitle column in the annotations.annotation list-column returned from nomis_data_info. For example, the content types returned from dataset "NM_1658_1", using nomis_data_info("NM_1658_1"), are "geoglevel", "2001census" and "sources".

nomis_content_type(content_type, id = NULL)

Arguments

content_type
A string with the content type to return metadata on.

id
A string with an optional content_type id.

Value

A tibble with metadata on a given content type.

See also

Examples

a <- nomis_content_type("sources")

tibble::glimpse(a)

#> Observations: 16
#> Variables: 6
#> $ description <chr> "A count of home Civil Service employees. It excludes t...
#> $ id          <chr> "acses", "aps", "ashe", "bres", "census", "cc", "dwp", ...
#> $ name        <chr> "Annual Civil Service Employment Survey", "Annual Popul...
#> $ type        <chr> "DataSource", "DataSource", "DataSource", "DataSource",...
#> $ item        <list> [NULL, NULL, NULL, <c("NA", "NA"), c("bres_curr", "bre...
#> $ status      <chr> NA, NA, NA, NA, NA, NA, "archived", "archived", NA, NA,...

b <- nomis_content_type("sources", id = "census")

tibble::glimpse(b)

#> Observations: 4
#> Variables: 9
#> $ description      <fct> Census data from the 1981, 1991, 2001 and 2011 Cen...
#> $ id               <fct> census, census, census, census
#> $ item.description <chr> "The 2011 Census was taken on 27th March 2011. The...
#> $ item.id          <chr> "census_2011", "census_2001", "census_1991", "cens...
#> $ item.item        <list> [<c("The Key Statistics series provides summary f...
#> $ item.name        <chr> "2011 Census", "2001 Census", "1991 Census", "1981...
#> $ item.type        <chr> "DataSourceSubGroup", "DataSourceSubGroup", "DataS...
#> $ name             <fct> Census of Population, Census of Population, Census...
#> $ type             <fct> DataSource, DataSource, DataSource, DataSource