Nomis data structures

Retrieve metadata on the structure and available variables for all available data sets or the information available in a specific dataset based on its ID.

nomis_data_info(id)

Arguments

id
Dataset ID. If empty, returns data on all available datasets. If the ID of a dataset, returns metadata for that particular dataset.

Value

A tibble with all available datasets and their metadata.

See also

Examples

# Get info on all datasets
x <- nomis_data_info()

tibble::glimpse(x)

#> Observations: 1,249
#> Variables: 14
#> $ agencyid                             <chr> "NOMIS", "NOMIS", "NOMIS", "NO...
#> $ id                                   <chr> "NM_1_1", "NM_2_1", "NM_4_1", ...
#> $ uri                                  <chr> "Nm-1d1", "Nm-2d1", "Nm-4d1", ...
#> $ version                              <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
#> $ annotations.annotation               <list> [<c("Current (being actively ...
#> $ components.attribute                 <list> [<c("Mandatory", "Conditional...
#> $ components.dimension                 <list> [<c("CL_1_1_GEOGRAPHY", "CL_1...
#> $ components.primarymeasure.conceptref <chr> "OBS_VALUE", "OBS_VALUE", "OBS...
#> $ components.timedimension.codelist    <chr> "CL_1_1_TIME", "CL_2_1_TIME", ...
#> $ components.timedimension.conceptref  <chr> "TIME", "TIME", "TIME", "TIME"...
#> $ description.value                    <chr> "Records the number of people ...
#> $ description.lang                     <chr> "en", "en", "en", "en", "en", ...
#> $ name.value                           <chr> "Jobseeker's Allowance with ra...
#> $ name.lang                            <chr> "en", "en", "en", "en", "en", ...

# Get info on a particular dataset
y <- nomis_data_info('NM_1658_1')

tibble::glimpse(y)

#> Observations: 1
#> Variables: 12
#> $ agencyid                             <chr> "NOMIS"
#> $ id                                   <chr> "NM_1658_1"
#> $ uri                                  <chr> "Nm-1658d1"
#> $ version                              <dbl> 1
#> $ annotations.annotation               <list> [<c("Historical (not actively...
#> $ components.attribute                 <list> [<c("Mandatory", "Conditional...
#> $ components.dimension                 <list> [<c("CL_1658_1_GEOGRAPHY", "C...
#> $ components.primarymeasure.conceptref <chr> "OBS_VALUE"
#> $ components.timedimension.codelist    <chr> "CL_1658_1_TIME"
#> $ components.timedimension.conceptref  <chr> "TIME"
#> $ name.value                           <chr> "UV035 - Distance travelled to...
#> $ name.lang                            <chr> "en"